Move JobConnections manager to its own file
